Diploma project. BlogEngine. Stage 3 Fixed #4.1
